feat: refactored to use individual files for components
