+++ src/ui/content.ts
@@ -1,7 +1,7 @@
 export const WATCH_FORM_HEADER = 'Watch any Ethereum account 👀';
 
 export const WATCH_FORM_DESCRIPTION =
-  'Enter any public address or *ENS name to add an account to watch within MetaMask.';
+  'Enter any public address or ENS name to add an account to watch within MetaMask.';
 
 export const WATCH_FORM_INSTRUCTIONS =
   'The watched accounts will be listed alongside the rest of your accounts in a safe, watch-only mode. Remember, you can look but you can’t sign or transact.';
@@ -10,6 +10,3 @@ export const WATCH_FORM_INPUT_LABEL = 'Ethereum address';
 
 export const WATCH_FORM_INPUT_PLACEHOLDER =
   'Enter a public address or ENS name';
-
-export const WATCH_FORM_ENS_DISCLAIMER =
-  '*ENS names are only supported on Ethereum mainnet';
